Split tests relying on sync queries vs unit tests
