+++ tests/features/multi_auth/test_add_operator_sync.py
